feat: move inMemory storages
