+++ src/main/java/ru/yandex/practicum/filmorate/storage/memory_old/InMemoryGenreStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.memory;
+package ru.yandex.practicum.filmorate.storage.memory_old;
 
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.storage.GenreStorage;
@@ -6,6 +6,16 @@ import ru.yandex.practicum.filmorate.storage.GenreStorage;
 import java.util.ArrayList;
 
 public class InMemoryGenreStorage implements GenreStorage {
+    @Override
+    public Genre addGenre(Genre genre) {
+        return null;
+    }
+
+    @Override
+    public Genre editGenre(Genre genre) {
+        return null;
+    }
+
     //TODO: заполнить
     public ArrayList<Genre> getAllGenres() {
         return null;
