+++ src/main/java/ru/yandex/practicum/filmorate/storage/memory_old/InMemoryRatingStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.memory;
+package ru.yandex.practicum.filmorate.storage.memory_old;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
